CHANGED: Iterating through arrays only once at hamming exercise.

+++ c/hamming/src/hamming.h
@@ -1,6 +1,6 @@
 #ifndef HAMMING_H
 #define HAMMING_H
 
-int compute(const char *lhs, const char *rhs);
+size_t compute(const char *lhs, const char *rhs);
 
 #endif
